fix(client): apply code-review fixes across currency, theme, and codegen
- Unify currency symbol resolution through AmountFormatter.getCurrencySymbol
  and remove remaining hard-coded defaults across report, finance, and profile
- Theme provider migrated to codegen-generated ThemeNotifier; drop manual
  theme_provider and associated overrides
- Remove dead shared_space notification provider in favor of a dedicated
  notification provider with extracted mutations
- Harden version checkUpdate to reset isChecking on error so retries are not
  stuck permanently; fix logging interceptor comments to match behavior
- Regenerate freezed/json/g files for changed models and providers

+++ client/lib/features/chat/genui/templates/transaction_group_receipt_sheets.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:forui/forui.dart';
+import 'package:decimal/decimal.dart';
 
 import 'package:finvo/app/theme/app_semantic_colors.dart';
 import 'package:finvo/i18n/strings.g.dart';
@@ -30,7 +31,7 @@ Widget _buildInfoRow(
 /// from the transaction's currency. Returns true when the user accepts.
 Future<bool> showCurrencyMismatchConfirmDialog(
   BuildContext context, {
-  required double amount,
+  required Decimal amount,
   required String fromCurrency,
   required String toCurrency,
   required String accountName,
